Set Edges does something now!
I renamed the discover method it "spawn_room" as that is something to
be handled by the map, in theory players will discover rooms. I'll add in room shuffling later, and make the
different types of rooms work later too.
The demo works again.
To start it just run ./tratior/main.py from the project directory.
The current options are go, and quit.