Modifications to how rooms handle direction.
Removed a faulty test.
set_edges provides a list of excepted values in the direction slot.
This list of accepted edges is trimmed down every time a connection is
made.

move changed to raise an assertion error if the direction give is in
the list of accepted edges.

This will be improved upon later.