Did some impovements to the demo
freak, I can't get the remove_from_available function to work.
For what ever reason, it likes to pretend that my strings are lists.
Despite specifically putting in a clause to ignore strings.
I suspect the problem is that it is deleting the shared state between
rooms rather than their own local copy.
But I am a bit to frustrated to look at the time being.
I'll look into setting up a copy function for the rooms rather than a
looking at the shared state.