Wrote Docs, Refactored
It occured to me, that at no point will there ever be more than one
map. In this case, the map class is redundant and doesn't offer any
signifigant benefit yet. I've removed it for now and have set the map
to exist in the house namespace. I also renamed housemap.py to house.py.
This is because all the functions are related to the house.