Finished the Room List
Emacs seems to have a stalling problem when the roomlist is put into
index.org as a result I will be moving the roomlist into rooms.org.
This shoukd reduce some of the strain while computing this.
I Also finished transcribing the room list into a literate programming
paradigm. Now in theory if I just have to add a property to the list,
and any changes should propagate automatically.