Figured out how to use dynamic blocks
I managed to get the roomlist to pass properly.
Now all you have to do if you want to make a new room is update
rooms.org and add it in. Then update the table on index.org by using
C-c C-c and tangling out.
Much easier and less room for failure.
Planning on adding in some additional neatness in the room.org file.