Map Roomlist objects! Reorganized Tests
I wrote a bit of documentation, I also moved around some of the
unittests. I'm planning on removing the move function from the rooms
all together. I made the roomlist an object and I made the map an
object. This is because globals are a pain to test, and I refuse to
deal with them again. Currently Map, Room, Roomlist tests are passing.
I'm going to just redo the entire discover fuction again because it
was not very good. run tests with: python3 -m unittest tests/*.py  at root